requirements.txt
# Core ML libraries
torch>=2.9.1
torchvision>=0.21.0
torchaudio>=2.9.1
transformers>=4.55.0
peft>=0.7.0
acceleratе>=0.25.0
bitsandbytes>=0.41.0
datasets>=2.16.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0
PyYAML>=6.0.0

# Metrics and evaluation
jiwer>=3.0.0
scikit-learn>=1.3.0

# Logging and monitoring
tensorboard>=2.15.0
tqdm>=4.65.0

# Utilities
huggingface-hub>=0.20.0
safetensors>=0.4.0
sentencepiece>=0.1.99

# Optional: Unsloth for faster LoRA training (если понадобится)
# unsloth @ git+https://github.com/unslothai/unsloth.git
